+++ src/Model/Customer.php
@@ -1,9 +1,10 @@
 <?php
-declare(strict_types=1);
 
 /**
  * Class CustomerInterface
  */
+
+declare(strict_types=1);
 
 namespace Paytrail\SDK\Model;
 
@@ -21,7 +22,6 @@ use Paytrail\SDK\Util\JsonSerializable;
  */
 class Customer implements \JsonSerializable, CustomerInterface
 {
-
     use JsonSerializable;
 
     /**
@@ -29,7 +29,7 @@ class Customer implements \JsonSerializable, CustomerInterface
      *
      * @throws ValidationException
      */
-    public function validate()
+    public function validate(): bool
     {
         $props = get_object_vars($this);
 
@@ -76,27 +76,32 @@ class Customer implements \JsonSerializable, CustomerInterface
     protected $vatId;
 
     /**
+     * The Company name
+     *
+     * @var string
+     */
+    protected $companyName;
+
+    /**
      * Get email.
      *
      * @return string
      */
     public function getEmail(): ?string
     {
-
         return $this->email;
     }
 
     /**
      * Set email.
      *
-     * @param string $email
+     * @param string|null $email
      *
      * @return self Return self to enable chaining.
      */
-    public function setEmail(?string $email) : CustomerInterface
+    public function setEmail(?string $email): CustomerInterface
     {
         $this->email = $email;
-
         return $this;
     }
 
@@ -107,21 +112,19 @@ class Customer implements \JsonSerializable, CustomerInterface
      */
     public function getFirstName(): ?string
     {
-
         return $this->firstName;
     }
 
     /**
      * Set first name.
      *
-     * @param string $firstName
+     * @param string|null $firstName
      *
      * @return self Return self to enable chaining.
      */
-    public function setFirstName(?string $firstName) : CustomerInterface
+    public function setFirstName(?string $firstName): CustomerInterface
     {
         $this->firstName = $firstName;
-
         return $this;
     }
 
@@ -132,21 +135,19 @@ class Customer implements \JsonSerializable, CustomerInterface
      */
     public function getLastName(): ?string
     {
-
         return $this->lastName;
     }
 
     /**
      * Set last name.
      *
-     * @param string $lastName
+     * @param string|null $lastName
      *
      * @return self Return self to enable chaining.
      */
-    public function setLastName(?string $lastName) : CustomerInterface
+    public function setLastName(?string $lastName): CustomerInterface
     {
         $this->lastName = $lastName;
-
         return $this;
     }
 
@@ -157,21 +158,19 @@ class Customer implements \JsonSerializable, CustomerInterface
      */
     public function getPhone(): ?string
     {
-
         return $this->phone;
     }
 
     /**
      * Set phone.
      *
-     * @param string $phone
+     * @param string|null $phone
      *
      * @return self Return self to enable chaining.
      */
-    public function setPhone(?string $phone) : CustomerInterface
+    public function setPhone(?string $phone): CustomerInterface
     {
         $this->phone = $phone;
-
         return $this;
     }
 
@@ -182,21 +181,42 @@ class Customer implements \JsonSerializable, CustomerInterface
      */
     public function getVatId(): ?string
     {
-
         return $this->vatId;
     }
 
     /**
      * Set VAT id.
      *
-     * @param string $vatId
+     * @param string|null $vatId
      *
      * @return self Return self to enable chaining.
      */
-    public function setVatId(?string $vatId) : CustomerInterface
+    public function setVatId(?string $vatId): CustomerInterface
     {
         $this->vatId = $vatId;
+        return $this;
+    }
 
+    /**
+     * Get Company name.
+     *
+     * @return string
+     */
+    public function getCompanyName(): ?string
+    {
+        return $this->companyName;
+    }
+
+    /**
+     * Set Company Name.
+     *
+     * @param string|null $companyName
+     *
+     * @return self Return self to enable chaining.
+     */
+    public function setCompanyName(?string $companyName): CustomerInterface
+    {
+        $this->companyName = $companyName;
         return $this;
     }
 }