+++ src/Model/Item.php
@@ -1,9 +1,10 @@
 <?php
-declare(strict_types=1);
 
 /**
  * Class ItemTest
  */
+
+declare(strict_types=1);
 
 namespace Paytrail\SDK\Model;
 
@@ -58,6 +59,7 @@ class Item implements \JsonSerializable, ItemInterface
      * The delivery date.
      *
      * @var string
+     * @deprecated
      */
     protected $deliveryDate;
 
@@ -96,7 +98,7 @@ class Item implements \JsonSerializable, ItemInterface
      * Merchant ID for the item.
      * Required for Shop-in-Shop payments, do not use for normal payments.
      *
-     * @var int
+     * @var string|null
      */
     protected $merchant;
 
@@ -124,7 +126,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param int $unitPrice
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setUnitPrice(?int $unitPrice) : ItemInterface
+    public function setUnitPrice(?int $unitPrice): ItemInterface
     {
         $this->unitPrice = $unitPrice;
 
@@ -147,7 +149,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param int $units
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setUnits(?int $units) : ItemInterface
+    public function setUnits(?int $units): ItemInterface
     {
         $this->units = $units;
 
@@ -170,7 +172,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param int $vatPercentage
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setVatPercentage(?int $vatPercentage) : ItemInterface
+    public function setVatPercentage(?int $vatPercentage): ItemInterface
     {
         $this->vatPercentage = $vatPercentage;
 
@@ -193,7 +195,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param string $productCode
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setProductCode(?string $productCode) : ItemInterface
+    public function setProductCode(?string $productCode): ItemInterface
     {
         $this->productCode = $productCode;
 
@@ -204,6 +206,8 @@ class Item implements \JsonSerializable, ItemInterface
      * Get the delivery date.
      *
      * @return string
+     *
+     * @deprecated Delivery date is deprecated
      */
     public function getDeliveryDate(): ?string
     {
@@ -215,8 +219,10 @@ class Item implements \JsonSerializable, ItemInterface
      *
      * @param string $deliveryDate
      * @return ItemInterface Return self to enable chaining.
+     *
+     * @deprecated Delivery date is deprecated
      */
-    public function setDeliveryDate(?string $deliveryDate) : ItemInterface
+    public function setDeliveryDate(?string $deliveryDate): ItemInterface
     {
         $this->deliveryDate = $deliveryDate;
 
@@ -239,7 +245,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param string $description
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setDescription(?string $description) : ItemInterface
+    public function setDescription(?string $description): ItemInterface
     {
         $this->description = $description;
 
@@ -262,7 +268,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param string $category
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setCategory(?string $category) : ItemInterface
+    public function setCategory(?string $category): ItemInterface
     {
         $this->category = $category;
 
@@ -285,7 +291,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param string $stamp
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setStamp(?string $stamp) : ItemInterface
+    public function setStamp(?string $stamp): ItemInterface
     {
         $this->stamp = $stamp;
 
@@ -308,7 +314,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param string $reference
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setReference(?string $reference) : ItemInterface
+    public function setReference(?string $reference): ItemInterface
     {
         $this->reference = $reference;
 
@@ -331,7 +337,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param string $merchant
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setMerchant(?string $merchant) : ItemInterface
+    public function setMerchant(?string $merchant): ItemInterface
     {
         $this->merchant = $merchant;
 
@@ -354,7 +360,7 @@ class Item implements \JsonSerializable, ItemInterface
      * @param CommissionInterface $commission
      * @return ItemInterface Return self to enable chaining.
      */
-    public function setCommission(?CommissionInterface $commission) : ItemInterface
+    public function setCommission(?CommissionInterface $commission): ItemInterface
     {
         $this->commission = $commission;
 
@@ -362,7 +368,7 @@ class Item implements \JsonSerializable, ItemInterface
     }
 
     /**
-     * Validates with Respect\Validation library and throws an exception for invalid objects
+     * Validates properties and throws an exception for invalid values
      *
      * @throws ValidationException
      */
@@ -370,19 +376,23 @@ class Item implements \JsonSerializable, ItemInterface
     {
         $props = get_object_vars($this);
 
+        if ($props['unitPrice'] === null) {
+            throw new ValidationException('Item unitPrice is empty');
+        }
         if ($props['unitPrice'] < 0) {
-            throw new ValidationException('Items UnitPrice can\'t be a negative number');
+            throw new ValidationException('Items unitPrice can\'t be a negative number');
         }
-
-        if (filter_var($props['unitPrice'], FILTER_VALIDATE_INT) === false) {
-            //throw new \Exception('UnitPrice is not an integer');
-            throw new ValidationException('UnitPrice is not an integer');
+        if ($props['unitPrice'] > 99999999) {
+            throw new ValidationException('Items unitPrice can\'t be over 99999999');
         }
-        if (filter_var($props['units'], FILTER_VALIDATE_INT) === false) {
-            throw new ValidationException('Units is not an integer');
+        if ($props['units'] === null) {
+            throw new ValidationException('Item units is empty');
         }
-        if (filter_var($props['vatPercentage'], FILTER_VALIDATE_INT) === false) {
-            throw new ValidationException('vatPercentage is not an integer');
+        if ($props['units'] < 0) {
+            throw new ValidationException('Items units can\'t be a negative number');
+        }
+        if ($props['vatPercentage'] < 0) {
+            throw new ValidationException('Items vatPercentage can\'t be a negative number');
         }
         if (empty($props['productCode'])) {
             throw new ValidationException('productCode is empty');
@@ -391,19 +401,19 @@ class Item implements \JsonSerializable, ItemInterface
         return true;
     }
 
-     /**
-      * Validates shop-in-shop props with Respect\Validation library and throws an exception for invalid objects
-      *
-      * @throws ValidationException
-      */
-     public function validateShopInShop()
-     {
-         $props = get_object_vars($this);
+    /**
+     * Validates shop-in-shop props and throws an exception for invalid values
+     *
+     * @throws ValidationException
+     */
+    public function validateShopInShop()
+    {
+        $props = get_object_vars($this);
 
-         if (empty($props['merchant'])) {
-             throw new ValidationException('merchant is empty');
-         }
+        if (empty($props['merchant'])) {
+            throw new ValidationException('merchant is empty');
+        }
 
-         return true;
-     }
+        return true;
+    }
 }