+++ src/PaytrailClient.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Paytrail\SDK;
@@ -44,20 +45,40 @@ abstract class PaytrailClient
     }
 
     /**
+     * A proxy for the Signature class' static method
+     * to be used via a client instance.
+     *
+     * @param array $response The response parameters.
+     * @param string $body The response body.
+     * @param string $signature The response signature key.
+     *
+     * @throws HmacException
+     */
+    abstract public function validateHmac(array $response = [], string $body = '', string $signature = '');
+
+    /**
      * A wrapper for post requests.
      *
      * @param string $uri The uri for the request.
-     * @param \JsonSerializable $data The request payload.
-     * @param callable|null $callback The callback method to run for the decoded response. If left empty, the response is returned.
-     * @param string|null $transactionId Paytrail transaction ID when accessing single transaction not required for a new payment request.
+     * @param \JsonSerializable|null $data The request payload.
+     * @param callable|null $callback The callback method to run for the decoded response.
+     * If left empty, the response is returned.
+     * @param string|null $transactionId Paytrail transaction ID when accessing single transaction.
+     * Not required for a new payment request.
      * @param bool $signatureInHeader Checks if signature is calculated from header/body parameters
      * @param string|null $paytrailTokenizationId Paytrail tokenization ID for getToken request
      *
      * @return mixed
      * @throws HmacException
      */
-    protected function post(string $uri, \JsonSerializable $data, callable $callback = null, string $transactionId = null, bool $signatureInHeader = true, string $paytrailTokenizationId = null)
-    {
+    protected function post(
+        string $uri,
+        \JsonSerializable $data = null,
+        callable $callback = null,
+        string $transactionId = null,
+        bool $signatureInHeader = true,
+        string $paytrailTokenizationId = null
+    ) {
         $body = json_encode($data, JSON_UNESCAPED_SLASHES);
 
         if ($signatureInHeader) {
@@ -77,8 +98,11 @@ abstract class PaytrailClient
             $headers = $this->reduceHeaders($response->getHeaders());
             $this->validateHmac($headers, $body, $headers['signature'] ?? '');
         } else {
+            // @phpstan-ignore-next-line FIXME
             $mac = $this->calculateHmac($data->toArray());
+            // @phpstan-ignore-next-line FIXME
             $data->setSignature($mac);
+            // @phpstan-ignore-next-line FIXME
             $body = json_encode($data->toArray(), JSON_UNESCAPED_SLASHES);
 
             $response = $this->http_client->request('POST', $uri, [
@@ -101,8 +125,10 @@ abstract class PaytrailClient
      * A wrapper for get requests.
      *
      * @param string $uri The uri for the request.
-     * @param callable|null $callback The callback method to run for the decoded response. If left empty, the response is returned.
-     * @param string|null $transactionId Paytrail transaction ID when accessing single transaction not required for a new payment request.
+     * @param callable|null $callback The callback method to run for the decoded response.
+     * If left empty, the response is returned.
+     * @param string|null $transactionId Paytrail transaction ID when accessing single transaction.
+     * Not required for a new payment request.
      *
      * @return mixed
      * @throws HmacException
@@ -136,13 +162,17 @@ abstract class PaytrailClient
      * Format request headers.
      *
      * @param string $method The request method. GET or POST.
-     * @param string|null $transactionId Paytrail transaction ID when accessing single transaction not required for a new payment request.
+     * @param string|null $transactionId Paytrail transaction ID when accessing single transaction.
+     * Not required for a new payment request.
      * @param string|null $checkoutTokenizationId Paytrail tokenization ID for getToken request
      *
      * @return array
      */
-    protected function getHeaders(string $method, string $transactionId = null, string $checkoutTokenizationId = null): array
-    {
+    protected function getHeaders(
+        string $method,
+        string $transactionId = null,
+        string $checkoutTokenizationId = null
+    ): array {
         $datetime = new \DateTime();
 
         $headers = [