+++ src/Request/MitPaymentRequest.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Class MitPaymentRequest
  */
@@ -23,7 +24,7 @@ class MitPaymentRequest extends AbstractPaymentRequest implements TokenPaymentRe
     protected $token;
 
     /**
-     * Validates with Respect\Validation library and throws an exception for invalid objects
+     * Validates properties and throws an exception for invalid values
      *
      * @throws ValidationException
      */