+++ src/Request/RefundRequest.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Class Refund
  */
@@ -19,11 +20,10 @@ use Paytrail\SDK\Util\JsonSerializable;
  */
 class RefundRequest implements \JsonSerializable
 {
-
     use JsonSerializable;
 
     /**
-     * Validates with Respect\Validation library and throws an exception for invalid objects
+     * Validates properties and throws an exception for invalid values
      *
      * @throws ValidationException
      */
@@ -31,8 +31,8 @@ class RefundRequest implements \JsonSerializable
     {
         $props = get_object_vars($this);
 
-        if (! empty($this->items)) {
-            // Count the total amount of the payment.
+        if (!empty($this->items)) {
+            // Count the total amount of the refund.
             $items_total = array_reduce($this->items, function ($carry = 0, ?RefundItem $item = null) {
                 if ($item === null) {
                     return $carry;
@@ -49,12 +49,8 @@ class RefundRequest implements \JsonSerializable
             $items_total = $this->amount;
         }
 
-        if (empty($props['amount'])) {
-            throw new ValidationException('Amount can not be empty');
-        }
-
-        if (filter_var($props['amount'], FILTER_VALIDATE_INT) === false) {
-            throw new ValidationException('Amount is not an integer');
+        if (empty($props['amount']) && empty($this->items)) {
+            throw new ValidationException('Amount can not be empty when making refund without items');
         }
 
         if ($items_total !== $props['amount']) {
@@ -65,6 +61,7 @@ class RefundRequest implements \JsonSerializable
             throw new ValidationException('CallbackUrls are not set');
         }
 
+        // Validate callbackUrls
         $this->callbackUrls->validate();
 
         return true;
@@ -94,11 +91,32 @@ class RefundRequest implements \JsonSerializable
     protected $callbackUrls;
 
     /**
+     * Refund recipient email address.
+     *
+     * @var string|null $email
+     */
+    protected $email;
+
+    /**
+     * Merchant unique identifier for the refund.
+     *
+     * @var string|null $refundStamp
+     */
+    protected $refundStamp;
+
+    /**
+     * Refund reference.
+     *
+     * @var string|null $refundReference
+     */
+    protected $refundReference;
+
+    /**
      * Get the amount.
      *
-     * @return int
+     * @return int|null
      */
-    public function getAmount() : int
+    public function getAmount(): ?int
     {
         return $this->amount;
     }
@@ -106,14 +124,13 @@ class RefundRequest implements \JsonSerializable
     /**
      * Set the amount.
      *
-     * @param int $amount
+     * @param int|null $amount
      *
      * @return RefundRequest Return self to enable chaining.
      */
-    public function setAmount(?int $amount) : RefundRequest
+    public function setAmount(?int $amount): RefundRequest
     {
         $this->amount = $amount;
-
         return $this;
     }
 
@@ -122,7 +139,7 @@ class RefundRequest implements \JsonSerializable
      *
      * @return RefundItem[]
      */
-    public function getItems() : array
+    public function getItems(): array
     {
         return $this->items ?? [];
     }
@@ -134,19 +151,18 @@ class RefundRequest implements \JsonSerializable
      *
      * @return RefundRequest Return self to enable chaining.
      */
-    public function setItems(?array $items) : RefundRequest
+    public function setItems(?array $items): RefundRequest
     {
         $this->items = $items;
-
         return $this;
     }
 
     /**
      * Get the callback urls.
      *
-     * @return CallbackUrl
+     * @return CallbackUrl|null
      */
-    public function getCallbackUrls() : CallbackUrl
+    public function getCallbackUrls(): ?CallbackUrl
     {
         return $this->callbackUrls;
     }
@@ -154,14 +170,79 @@ class RefundRequest implements \JsonSerializable
     /**
      * Set the callback urls.
      *
-     * @param CallbackUrl $callbackUrls The callback url instance holding success and cancel urls.
+     * @param CallbackUrl|null $callbackUrls The callback url instance holding success and cancel urls.
      *
      * @return RefundRequest Return self to enable chaining.
      */
-    public function setCallbackUrls(?CallbackUrl $callbackUrls) : RefundRequest
+    public function setCallbackUrls(?CallbackUrl $callbackUrls): RefundRequest
     {
         $this->callbackUrls = $callbackUrls;
+        return $this;
+    }
 
+    /**
+     * Get customer email.
+     *
+     * @return string|null
+     */
+    public function getEmail(): ?string
+    {
+        return $this->email;
+    }
+
+    /**
+     * Set customer email
+     *
+     * @param string|null $email
+     * @return RefundRequest
+     */
+    public function setEmail(?string $email): RefundRequest
+    {
+        $this->email = $email;
+        return $this;
+    }
+
+    /**
+     * Get refund stamp.
+     *
+     * @return string|null
+     */
+    public function getRefundStamp(): ?string
+    {
+        return $this->refundStamp;
+    }
+
+    /**
+     * Set refund stamp.
+     *
+     * @param string|null $refundStamp
+     * @return RefundRequest
+     */
+    public function setRefundStamp(?string $refundStamp): RefundRequest
+    {
+        $this->refundStamp = $refundStamp;
+        return $this;
+    }
+
+    /**
+     * Get refund reference.
+     *
+     * @return string|null
+     */
+    public function getRefundReference(): ?string
+    {
+        return $this->refundReference;
+    }
+
+    /**
+     * Set refund reference.
+     *
+     * @param string|null $refundReference
+     * @return RefundRequest
+     */
+    public function setRefundReference(?string $refundReference): RefundRequest
+    {
+        $this->refundReference = $refundReference;
         return $this;
     }
 }