+++ src/Request/RevertPaymentAuthHoldRequest.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Class RevertCitPaymentAuthHoldRequest
  */
@@ -25,7 +26,7 @@ class RevertPaymentAuthHoldRequest implements \JsonSerializable
     protected $transactionId;
 
     /**
-     * Validates with Respect\Validation library and throws an exception for invalid objects
+     * Validates properties and throws an exception for invalid values
      *
      * @throws ValidationException
      */