+++ src/Response/CitPaymentResponse.php
@@ -1,9 +1,10 @@
 <?php
-declare(strict_types=1);
 
 /**
  * Class CitPaymentResponse
  */
+
+declare(strict_types=1);
 
 namespace Paytrail\SDK\Response;
 
@@ -31,14 +32,13 @@ class CitPaymentResponse implements ResponseInterface
     /**
      * Set the transaction id.
      *
-     * @param string $transactionId
+     * @param string|null $transactionId
      *
      * @return CitPaymentResponse Return self to enable chaining.
      */
-    public function setTransactionId(string $transactionId): CitPaymentResponse
+    public function setTransactionId(?string $transactionId): CitPaymentResponse
     {
         $this->transactionId = $transactionId;
-
         return $this;
     }
 
@@ -53,7 +53,7 @@ class CitPaymentResponse implements ResponseInterface
     }
 
     /**
-     * @param string $threeDSecureUrl
+     * @param string|null $threeDSecureUrl
      * @return CitPaymentResponse
      */
     public function setThreeDSecureUrl(?string $threeDSecureUrl): CitPaymentResponse