+++ src/Response/MitPaymentResponse.php
@@ -1,9 +1,10 @@
 <?php
-declare(strict_types=1);
 
 /**
  * Class MitPaymentResponse
  */
+
+declare(strict_types=1);
 
 namespace Paytrail\SDK\Response;
 
@@ -28,14 +29,13 @@ class MitPaymentResponse implements ResponseInterface
     /**
      * Set the transaction id.
      *
-     * @param string $transactionId
+     * @param string|null $transactionId
      *
      * @return MitPaymentResponse Return self to enable chaining.
      */
-    public function setTransactionId(string $transactionId): MitPaymentResponse
+    public function setTransactionId(?string $transactionId): MitPaymentResponse
     {
         $this->transactionId = $transactionId;
-
         return $this;
     }