+++ src/Response/PaymentStatusResponse.php
@@ -1,9 +1,10 @@
 <?php
-declare(strict_types=1);
 
 /**
  * Class PaymentResponse
  */
+
+declare(strict_types=1);
 
 namespace Paytrail\SDK\Response;
 
@@ -35,7 +36,7 @@ class PaymentStatusResponse implements ResponseInterface
 
     /**
      * Total amount of the payment in currency's minor units,
-     * eg. for Euros means cents.
+     * e.g. for Euros means cents.
      *
      * @var integer
      */
@@ -104,7 +105,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get the transaction id.
      *
      * @return string
-     */ 
+     */
     public function getTransactionId(): string
     {
         return $this->transactionId;
@@ -116,7 +117,7 @@ class PaymentStatusResponse implements ResponseInterface
      * @param string $transactionId
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setTransactionId(string $transactionId): PaymentStatusResponse
     {
         $this->transactionId = $transactionId;
@@ -128,7 +129,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get the payment status.
      *
      * @return string
-     */ 
+     */
     public function getStatus(): ?string
     {
         return $this->status;
@@ -137,15 +138,14 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set the payment status.
      *
-     * @param string $status Possible values: new, ok, fail,
+     * @param string|null $status Possible values: new, ok, fail,
      * pending, or delayed
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setStatus(?string $status): PaymentStatusResponse
     {
         $this->status = $status;
-
         return $this;
     }
 
@@ -153,7 +153,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get the total amount of the payment.
      *
      * @return integer
-     */ 
+     */
     public function getAmount(): ?int
     {
         return $this->amount;
@@ -161,16 +161,15 @@ class PaymentStatusResponse implements ResponseInterface
 
     /**
      * Set the total amount of the payment in currency's minor units,
-     * eg. for Euros means cents.
+     * e.g. for Euros means cents.
      *
-     * @param integer $amount
+     * @param integer|null $amount
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setAmount(?int $amount): PaymentStatusResponse
     {
         $this->amount = $amount;
-
         return $this;
     }
 
@@ -178,7 +177,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get currency.
      *
      * @return string
-     */ 
+     */
     public function getCurrency(): ?string
     {
         return $this->currency;
@@ -187,14 +186,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set currency.
      *
-     * @param string $currency
+     * @param string|null $currency
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setCurrency(?string $currency): PaymentStatusResponse
     {
         $this->currency = $currency;
-
         return $this;
     }
 
@@ -202,7 +200,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get merchant unique identifier for the order.
      *
      * @return string
-     */ 
+     */
     public function getStamp(): ?string
     {
         return $this->stamp;
@@ -211,14 +209,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set merchant unique identifier for the order.
      *
-     * @param string $stamp
+     * @param string|null $stamp
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setStamp(?string $stamp): PaymentStatusResponse
     {
         $this->stamp = $stamp;
-
         return $this;
     }
 
@@ -226,8 +223,8 @@ class PaymentStatusResponse implements ResponseInterface
      * Get the order reference.
      *
      * @return string
-     */ 
-    public function getReference()
+     */
+    public function getReference(): string
     {
         return $this->reference;
     }
@@ -235,14 +232,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set the order reference.
      *
-     * @param string $reference
+     * @param string|null $reference
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setReference(?string $reference): PaymentStatusResponse
     {
         $this->reference = $reference;
-
         return $this;
     }
 
@@ -250,7 +246,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get the transaction creation timestamp.
      *
      * @return string
-     */ 
+     */
     public function getCreatedAt(): ?string
     {
         return $this->createdAt;
@@ -259,14 +255,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set the transaction creation timestamp.
      *
-     * @param string $createdAt
+     * @param string|null $createdAt
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setCreatedAt(?string $createdAt): PaymentStatusResponse
     {
         $this->createdAt = $createdAt;
-
         return $this;
     }
 
@@ -274,7 +269,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get payment API url.
      *
      * @return string
-     */ 
+     */
     public function getHref(): ?string
     {
         return $this->href;
@@ -283,14 +278,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set payment API url.
      *
-     * @param string $href
+     * @param string|null $href
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setHref(?string $href): PaymentStatusResponse
     {
         $this->href = $href;
-
         return $this;
     }
 
@@ -298,7 +292,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get provider name.
      *
      * @return string
-     */ 
+     */
     public function getProvider()
     {
         return $this->provider;
@@ -307,14 +301,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set provider name.
      *
-     * @param string $provider If processed, the name of the payment method provider
+     * @param string|null $provider If processed, the name of the payment method provider
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setProvider(?string $provider): PaymentStatusResponse
     {
         $this->provider = $provider;
-
         return $this;
     }
 
@@ -322,7 +315,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get filing code.
      *
      * @return string
-     */ 
+     */
     public function getFilingCode(): ?string
     {
         return $this->filingCode;
@@ -331,14 +324,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set filing code.
      *
-     * @param string $filingCode
+     * @param string|null $filingCode
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setFilingCode(?string $filingCode): PaymentStatusResponse
     {
         $this->filingCode = $filingCode;
-
         return $this;
     }
 
@@ -346,7 +338,7 @@ class PaymentStatusResponse implements ResponseInterface
      * Get timestamp when the transaction was paid
      *
      * @return string
-     */ 
+     */
     public function getPaidAt(): ?string
     {
         return $this->paidAt;
@@ -355,14 +347,13 @@ class PaymentStatusResponse implements ResponseInterface
     /**
      * Set timestamp when the transaction was paid
      *
-     * @param string $paidAt
+     * @param string|null $paidAt
      *
      * @return PaymentStatusResponse Return self to enable chaining.
-     */ 
+     */
     public function setPaidAt(?string $paidAt): PaymentStatusResponse
     {
         $this->paidAt = $paidAt;
-
         return $this;
     }
 }