+++ tests/Model/ItemTest.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Tests\Model;
@@ -11,83 +12,103 @@ class ItemTest extends TestCase
 {
     public function testIsItemValid()
     {
-        $i = new Item;
-
-        $this->assertInstanceOf(
-            Item::class,
-            $i
-        );
-
-        $i->setUnitPrice(1)
+        $item = (new Item())->setUnitPrice(1)
             ->setUnits(2)
+            ->setStamp('thisIsStamp')
             ->setVatPercentage(0)
             ->setProductCode('productCode123')
-            ->setDeliveryDate('12.12.1999')
+            ->setDeliveryDate('2023-01-01')
             ->setDescription('description');
 
-        $this->assertEquals(true, $i->validate());
+        $this->assertTrue($item->validate());
     }
 
-    public function testExceptions()
+    public function testItemWithoutUnitPriceThrowsError()
     {
         $this->expectException(ValidationException::class);
-        $i = new Item;
-        $i->validate();
+        (new Item())->setUnits(2)
+            ->setStamp('thisIsStamp')
+            ->setVatPercentage(0)
+            ->setProductCode('productCode123')
+            ->setDescription('description')
+            ->validate();
     }
 
-    public function testExceptionMessages()
+    public function testItemWithoutUnitsTrowsError()
     {
-        $i = new Item;
-
-        try {
-            $i->validate();
-        } catch (\Exception $e) {
-            $this->assertStringContainsString('UnitPrice is not an integer', $e->getMessage());
-        }
-
-        try {
-            $i->setUnitPrice(-10);
-            $i->validate();
-        } catch (\Exception $e) {
-            $this->assertStringContainsString('UnitPrice can\'t be a negative number', $e->getMessage());
-        }
-
-        try {
-            $i->setUnitPrice(2);
-            $i->validate();
-        } catch (\Exception $e) {
-            $this->assertStringContainsString('Units is not an integer', $e->getMessage());
-        }
-
-        try {
-            $i->setUnits(3);
-            $i->validate();
-        } catch (ValidationException $e) {
-            $this->assertStringContainsString('vatPercentage is not an integer', $e->getMessage());
-        }
-
-        try {
-            $i->setVatPercentage(12);
-            $i->validate();
-        } catch (ValidationException $e) {
-            $this->assertStringContainsString('productCode is empty', $e->getMessage());
-        }
-
-        try {
-            $i->setProductCode('productCode123');
-            $i->validate();
-        } catch (ValidationException $e) {
-            $this->assertStringContainsString('deliveryDate is empty', $e->getMessage());
-        }
-
-        $i->setDeliveryDate('12.12.1999');
-
-        try {
-            $this->assertIsBool($i->validate(), 'Item::validate returns bool');
-        } catch (ValidationException $e) {
-        }
-
-
+        $this->expectException(ValidationException::class);
+        (new Item())->setUnitPrice(1)
+            ->setStamp('thisIsStamp')
+            ->setVatPercentage(0)
+            ->setProductCode('productCode123')
+            ->setDescription('description')
+            ->validate();
     }
 
+    public function testItemWithNegativeUnitsThrowsError()
+    {
+        $this->expectException(ValidationException::class);
+        (new Item())->setUnitPrice(1)
+            ->setUnits(-1)
+            ->setStamp('thisIsStamp')
+            ->setVatPercentage(0)
+            ->setProductCode('productCode123')
+            ->setDescription('description')
+            ->validate();
+    }
+
+    public function testItemWIthNegativeVatThrowsError()
+    {
+        $this->expectException(ValidationException::class);
+        (new Item())->setUnitPrice(1)
+            ->setUnits(2)
+            ->setStamp('thisIsStamp')
+            ->setVatPercentage(-10)
+            ->setProductCode('productCode123')
+            ->setDescription('description')
+            ->validate();
+    }
+
+    public function testItemWithoutProductCodeThrowsError()
+    {
+        $this->expectException(ValidationException::class);
+        (new Item())->setUnitPrice(1)
+            ->setUnits(2)
+            ->setStamp('thisIsStamp')
+            ->setVatPercentage(10)
+            ->setDescription('description')
+            ->validate();
+    }
+
+    public static function providerForUnitPriceLimitValues()
+    {
+        return [
+            'Negative amount' => [-1, false],
+            'Zero amount' => [0, true],
+            'Maximum amount' => [99999999, true],
+            'Over maximum amount' => [100000000, false]
+        ];
+    }
+
+    /**
+     * @dataProvider providerForUnitPriceLimitValues
+     */
+    public function testUnitPriceLimitValues($unitPrice, $expectedResult)
+    {
+        $item = (new Item())->setUnitPrice($unitPrice)
+            ->setUnits(2)
+            ->setStamp('thisIsStamp')
+            ->setVatPercentage(0)
+            ->setProductCode('productCode123')
+            ->setDeliveryDate('2023-01-01')
+            ->setDescription('description');
+
+        try {
+            $validationResult = $item->validate();
+        } catch (ValidationException $exception) {
+            $validationResult = false;
+        }
+
+        $this->assertEquals($expectedResult, $validationResult);
+    }
 }