+++ tests/Request/AddCardFormRequestTest.php
@@ -1,12 +1,6 @@
 <?php
-declare(strict_types=1);
 
-/**
- * Created by PhpStorm.
- * User: kimmok
- * Date: 03/04/2020
- * Time: 9.15
- */
+declare(strict_types=1);
 
 namespace Tests\Request;
 
@@ -16,7 +10,7 @@ use PHPUnit\Framework\TestCase;
 
 class AddCardFormRequestTest extends TestCase
 {
-    public function validationProvider()
+    public static function validationProvider()
     {
         return [
             'checkout-account is empty' => [